Fixing route switch bug
The issue was when switching routes, the ScrollToTop component works but only some milliseconds after the children are painted. Thus there was a 'glitch' seeing the content before the scroll was effected. After numerous attempts to solve this by reducing that time to near-zero, I gave up and decided to make a HOC to simply hide the entire app at #root for 0 seconds when any route is loaded. After 0 seconds (that is, at the end of the route change and consequent code execution, the #root is returned to 1 opacity.